requirements.txt
aiofiles==23.2.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.3.0
autopep8==2.0.4
billiard==4.2.0
black==24.1.1
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2
django-allauth==0.59.0
django-cors-headers==4.3.1
django-rest-passwordreset==1.2.1
django-silk==5.0.4
djangorestframework==3.15.1
dnspython==2.6.1
drf-jwt==1.19.2
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
ecdsa==0.18.0
exceptiongroup==1.2.0
gprof2dot==2022.7.29
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
model-bakery==1.17.0
mongoengine==0.28.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.2
pytest==7.4.3
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.15.2
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
